+++ src/lib/Block/Block.stories.svelte
@@ -2,21 +2,21 @@
   import Block from "./Block.svelte";
 
   export const meta = {
-    title: "Layout/Block",
+    title: "Components/Block",
     description: "A basic building block of a page",
     component: Block,
     tags: ["autodocs"],
     argTypes: {
       width: {
         default: "body",
-        options: ["body", "full"],
+        options: ["body", "wide", "full"],
         control: "select"
       }
     },
     parameters: {
       docs: {
         description: {
-          component: "A basic content block."
+          component: "A basic content block with several width options. This helps when building a page layout if you'd like to place your own components inside a container that aligns with the body well of a typical urban.org layout, a wider block, or a full width block."
         }
       }
     }