+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Next
 
 - Add Analytics component
+- Move Block documentation to main component section
 - Adds Basic page layout example
+- Update Storybook favicon to Urban logo
 
 ## v0.7.0